refactor(user): 관리자 사용자 목록 모델 분리
도메인 분리
- UserAdminViewService 내부 record였던 AdminUserRow와 UserFilter를 별도 파일로 분리했습니다.
- UserViewController는 서비스 nested type import 대신 독립 dto 패키지를 참조합니다.

의도
- 서비스 파일은 조회/명령 위임에 집중하고, 화면 모델 계약은 파일 단위로 추적되게 했습니다.

Refs #139
